Exp-1/output/f00874_transcribe_audio_test.py
from f00874_transcribe_audio import *
def test_transcribe_audio():
    audio_data = torch.tensor([0.1, 0.2, 0.3, 0.4])
    model = Wav2Vec2ForCTC()
    processor = Wav2Vec2Processor()

    transcription = transcribe_audio(audio_data, model, processor)

    assert transcription == 'transcribed_text'

    audio_data = torch.tensor([0.5, 0.6, 0.7, 0.8])

    transcription = transcribe_audio(audio_data, model, processor)

    assert transcription == 'transcribed_text'

    audio_data = torch.tensor([0.9, 1.0, 1.1, 1.2])

    transcription = transcribe_audio(audio_data, model, processor)

    assert transcription == 'transcribed_text'

    print('All tests passed!')


test_transcribe_audio()

